Made some necessary changes
- The MainMenu.cs script wasn't in the Scripts folder, it was just floating in Assets. I put it in Scripts. Please be more careful about this.
