+++ CompSciClubSpring2018/Assets/2DArt.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 6634e887b8a71f34e9535cd8db59a6cc
+folderAsset: yes
+timeCreated: 1518836913
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
