+++ CompSciClubSpring2018/Assets/3DArt.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 249626c58ac2666449ba9ef056203d20
+guid: 300c9d6d15293924299dc54571c8a0b6
 folderAsset: yes
-timeCreated: 1520710371
+timeCreated: 1518836913
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
