--- CompSciClubSpring2018/Assets/MainMenu.cs
@@ -1,19 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-public class MainMenu : MonoBehaviour {
-
-    public void PlayGame()
-    {
-        SceneManager.LoadScene("HunterGoodin-FirstScene");
-        Time.timeScale = 1f;
-    }
-
-    public void QuitGame()
-    {
-        Debug.Log("QUIT!");
-        Application.Quit();
-    }
-}
