+++ CompSciClubSpring2018/Assets/Materials.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: d67d688724365d941bc4a6fff5649e88
-timeCreated: 1520385984
+guid: b88204fc1a9ba584a99af318f7647fd8
+folderAsset: yes
+timeCreated: 1518837207
 licenseType: Free
 DefaultImporter:
   externalObjects: {}
