+++ CompSciClubSpring2018/Assets/Scripts/Pause Menu.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 16a9be2ae34225846b37e0ed386c418b
+folderAsset: yes
+timeCreated: 1522180214
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
