+++ CompSciClubSpring2018/Assets/StreamingAssets.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 2de65270794bc5e4a80220de873eedcb
+folderAsset: yes
+timeCreated: 1522178503
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
