+++ CompSciClubSpring2018/Assets/Textures.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: c9371ffffe94a20428639e0242205dd4
+folderAsset: yes
+timeCreated: 1518837207
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
